pyproject.toml
[project]
name = "telex-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.120.2",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "openai>=2.6.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
